qa: Test both Overwinter and Sapling peer management
