+++ qa/rpc-tests/test_framework/mininode.py
@@ -42,6 +42,7 @@ from .equihash import (
 OVERWINTER_PROTO_VERSION = 170003
 BIP0031_VERSION = 60000
 SPROUT_PROTO_VERSION = 170002  # past bip-31 for ping/pong
+SAPLING_PROTO_VERSION = 170006
 MY_SUBVERSION = "/python-mininode-tester:0.0.1/"
 
 OVERWINTER_VERSION_GROUP_ID = 0x03C48270
@@ -1415,7 +1416,7 @@ class NodeConn(asyncore.dispatcher):
         vt.addrFrom.port = 0
         self.send_message(vt, True)
         print 'MiniNode: Connecting to Bitcoin Node IP # ' + dstaddr + ':' \
-            + str(dstport)
+            + str(dstport) + ' using version ' + str(protocol_version)
 
         try:
             self.connect((dstaddr, dstport))
